Milestone 3 reached
Stage set up for a continuous delivery workflow (db-migrate)
