+++ package.json
@@ -12,6 +12,7 @@
     "async": "^1.5.0",
     "bower": "^1.7.1",
     "db-migrate": "^0.10.0-beta.6",
+    "db-migrate-sqlite3": "^0.1.3",
     "mysql": "^2.9.0",
     "node-dbi": "^0.7.1",
     "percolator": "^1.5.0"
