+++ src/backend/index.js
@@ -1,27 +1,9 @@
 'use strict';
 
-var Percolator = require('percolator').Percolator;
-var dbSession = require('../../src/backend/dbSession.js');
+var Server = require('./server.js').Server;
 
-var port = 8080;
-var server = Percolator({'port': port, 'autoLink': false});
-
-server.route('/api/keywords',
-{
-	GET: function (req, res) {
-		dbSession.fetchAll('SELECT id, value, categoryID FROM keyword ORDER BY id',
-			function(err, rows) {
-				if (err) {
-					console.log(err);
-					res.status.internalServerError(err);
-				} else {
-					res.collection(rows).send();
-				}
-			});
-	}
-}
-);
+var server = Server('8080');
 
 server.listen(function() {
-	console.log('Server started and listening on port', port);
+	console.log('Server started and listening on port', server.options.port);
 });
